# packages/back/built/api/api.yaml
openapi: 3.0.1
info:
    title: Tennis APP  API
    description: This service provides tennis players statics.
    contact:
        name: Slim HAMDI
        email: sl.hamdi.slim@gmail.com
    version: 1.0.0
servers:
    - url: /api/v1
      description: Current Server
    - url: http://localhost:5000/api/v1
      description: Production
paths:
...
    /player/{playerId}:
        get:
            summary: Returns the player requested for a customer paginated.
            description: Fetches  players for a specific player by id
            security:
                - authzero: []
            operationId: getPlayer
            tags:
                - customer
            parameters:
                - name: playerId
                  in: path
                  required: true
                  description: The id of the player.
                  schema:
                      type: number
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Player'
                400:
                    description: Bad request - The caller has provided a pagination header with bad boundaries
                401:
                    description: Caller did not have a valid token
                403:
                    description: Caller has a valid token but it not permitted to call this operation
                503:
                    description: Service or a dependency is unhealthy
...
components:
    schemas:
...
        Player:
            type: object
            properties:
                id:
                    type: number
                    description: the id of the player
                firstname:
                    type: string
                    description: firstname of the requested player
                lastname:
                    type: string
                    description: lastname of the requested player
                shortname:
                    type: string
                    description: shortname of the requested player
                sex:
                    type: string
                    description: sex of the requested player M or F
                country:
                    type: object
                    properties:
                        picture:
                            type: string
                            description: picture of the player country
                        code:
                            type: string
                            description: country code
                picture:
                    type: string
                    description: player picture
                data:
                    type: object
                    properties:
                        rank:
                            type: number
                            description: player rank
                        points:
                            type: number
                            description: number of collected points for the player
                        weight:
                            type: number
                            description: player weight
                        height:
                            type: number
                            description: player height
                        age:
                            type: number
                            description: player age
                        last:
                            type: array
                            description: array last sets
                            items:
                                type: number
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer